requirements.txt
fastapi
uvicorn

streamlit

faiss-cpu

google-generativeai
langchain

python-dotenv
requests
pandas

sentence-transformers
torch
langchain-huggingface
langchain
bitsandbytes
transformers